requirements.txt
requests
python-dotenv
pyyaml
jinja2